Changed pretty much everything. Really trying to get actual interfacing done. Some of it is probably useless since I am very tired.
If anyone is trying to track how things work:
MainExec runs a loop of displaying the menu and waiting for input.
Input goes into CLIDisplay where submenu logic is done(such as editing things), and passed to Controller.
Controller should distribute as necessary(selecting, deleting, sorting, etc.)
Once logic is done, it returns to Display which returns to MainExec, which loops

Added TODO's wherever I found things that didn't seem right, or need more thought.
Please go through them and if you find something with your name on it, it is because I thought that you would be the best equipped to do it.

I'll see you in a few hours.

Finally: Sorry for the commit message Wesley.